+++ cmdline/CtaAdminCmdParse.hpp
@@ -266,7 +266,8 @@ const std::map<std::string, OptionBoolean::Key> boolOptions = {
    { "--justretrieve",          OptionBoolean::JUSTRETRIEVE },
    { "--log",                   OptionBoolean::SHOW_LOG_ENTRIES },
    { "--lookupnamespace",       OptionBoolean::LOOKUP_NAMESPACE },
-   { "--summary",               OptionBoolean::SUMMARY }
+   { "--summary",               OptionBoolean::SUMMARY },
+   { "--no-recall", OptionBoolean::NO_RECALL}
 };
 
 
@@ -378,7 +379,8 @@ const std::map<AdminCmd::Cmd, CmdHelp> cmdHelp = {
 			   "     By default, CTA will migrate AND add new (or missing) copies (as defined by the storage class) of the files located on the tape to repack.\n"
                            "   * The --mountpolicy option allows to give a specific mount policy that will be applied to the repack subrequests (retrieve and archive requests).\n"
 			   "     By default, a hardcoded mount policy is applied (every request priorities and minimum request ages = 1).\n"
-			   "   * If the --disabledtape flag is set, the tape to repack will be mounted for retrieval even if it is disabled."
+			   "   * If the --disabledtape flag is set, the tape to repack will be mounted for retrieval even if it is disabled.\n"
+         "   * If the --no-recall flag is set, no retrieve mount will be triggered and only the files that are located in the buffer will be considered for archival."
 					"\n\n" 
 					 }},
    { AdminCmd::CMD_REQUESTERMOUNTRULE,   { "requestermountrule",   "rmr", { "add", "ch", "rm", "ls" } }},
@@ -501,6 +503,7 @@ const Option opt_refresh_interval     { Option::OPT_UINT,  "--refreshinterval", 
 const Option opt_targeted_free_space  { Option::OPT_UINT,  "--targetedfreespace",    "-f", " <targeted_free_space>" };
 const Option opt_sleep_time           { Option::OPT_UINT,  "--sleeptime",            "-s", " <sleep time in s>" };
 const Option opt_reason               { Option::OPT_STR,   "--reason",               "-r", " <reason_status_change>" };
+const Option opt_no_recall            { Option::OPT_FLAG, "--no-recall", "-nr", "" };
 
 /*!
  * Map valid options to commands
@@ -575,7 +578,7 @@ const std::map<cmd_key_t, cmd_val_t> cmdOptions = {
    {{ AdminCmd::CMD_MOUNTPOLICY,          AdminCmd::SUBCMD_LS    }, { }},
    /*----------------------------------------------------------------------------------------------------*/
    {{ AdminCmd::CMD_REPACK,               AdminCmd::SUBCMD_ADD   },
-      { opt_vid.optional(), opt_vidfile.optional(), opt_bufferurl.optional(), opt_justmove.optional(), opt_justaddcopies.optional(), opt_mountpolicy, opt_disabled_tape.optional() }},
+      { opt_vid.optional(), opt_vidfile.optional(), opt_bufferurl.optional(), opt_justmove.optional(), opt_justaddcopies.optional(), opt_mountpolicy, opt_disabled_tape.optional(), opt_no_recall.optional() }},
    {{ AdminCmd::CMD_REPACK,               AdminCmd::SUBCMD_RM    }, { opt_vid }},
    {{ AdminCmd::CMD_REPACK,               AdminCmd::SUBCMD_LS    }, { opt_vid.optional() }},
    {{ AdminCmd::CMD_REPACK,               AdminCmd::SUBCMD_ERR   }, { opt_vid }},