+++ common/utils/utils.hpp
@@ -462,6 +462,15 @@ namespace utils {
     return ret;
   }
   
+   /**
+   * Allows to remove the first c chars from the beginning of a string
+   * Example : removePrefix(aaaaaabcdef,'a') -> bcdef
+   * @param input the string from which we will remove the chars c from it
+   * @param prefixChar the character that we want to remove from the beginning
+   * @return a new string with the prefix.
+   */
+  std::string removePrefix(const std::string & input, char prefixChar);
+  
 } // namespace utils
 
 } // namespace cta