+++ xroot_plugins/XrdSsiCtaRequestMessage.cpp
@@ -1487,10 +1487,13 @@ void RequestMessage::processRepack_Add(cta::xrd::Response &response)
   }
 
   bool forceDisabledTape = has_flag(OptionBoolean::DISABLED);
+  
+  bool noRecall = has_flag(OptionBoolean::NO_RECALL);
 
   // Process each item in the list
   for(auto it = vid_list.begin(); it != vid_list.end(); ++it) {
-     m_scheduler.queueRepack(m_cliIdentity, *it, bufferURL, type, mountPolicy, forceDisabledTape, m_lc);
+    SchedulerDatabase::QueueRepackRequest repackRequest(*it,bufferURL,type,mountPolicy,forceDisabledTape, noRecall);
+    m_scheduler.queueRepack(m_cliIdentity, repackRequest, m_lc);
   }
 
   response.set_type(cta::xrd::Response::RSP_SUCCESS);